extended sdss bal mask to also work with desi data and linear binning, fixes to dla_mask
